# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info":{
        "title": "Inkster",
        "summary": "illustrations",
        "description": "Inkster is a platform for users to share, discover, and interact with illustrations, novels, and manga. This API provides endpoints for authentication, post management, reporting, and user interactions."    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/map/nearby-exhibitions": {
            "get": {
                "tags": ["map"],
                "summary": "Get nearby exhibitions",
                "description": "Fetches nearby art galleries and museums using Google Places API based on user's current coordinates.",
                "security": [
                {
                    "bearerAuth": []
                }
                ],
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude of the user's location",
                        "required": true,
                        "schema": {
                        "type": "string",
                        "example": "48.8566"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitude of the user's location",
                        "required": true,
                        "schema": {
                        "type": "string",
                        "example": "2.3522"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Search radius in meters (default: 5000)",
                        "required": false,
                        "schema": {
                        "type": "integer",
                        "example": 5000
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of nearby exhibitions",
                        "content": {
                        "application/json": {
                            "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Louvre Museum"
                                },
                                "place_id": {
                                    "type": "string",
                                    "example": "ChIJb8Jg9pJU5kcR7GBcDHgV6bs"
                                },
                                "address": {
                                    "type": "string",
                                    "example": "Rue de Rivoli, 75001 Paris, France"
                                },
                                "type": {
                                    "type": "string",
                                    "example": "museum"
                                },
                                "location": {
                                    "type": "object",
                                    "properties": {
                                    "lat": {
                                        "type": "number",
                                        "example": 48.8606
                                    },
                                    "lng": {
                                        "type": "number",
                                        "example": 2.3376
                                    }
                                    }
                                }
                                }
                            }
                            }
                        }
                        }
                    },
                    "400": {
                        "description": "Missing or invalid latitude or longitude"
                    },
                    "500": {
                        "description": "Server error or external API failure"
                    }
                }
            }
        },
...
    }
}